refactor: Use CSS Modules for MailingListSignup.
