+++ components/Layout.tsx
@@ -1,7 +1,7 @@
 import Head from 'next/head';
 import { Header } from './Header';
 import { Footer } from './Footer/Footer';
-import { MailingListSignupForm } from './MailingListSignup';
+import { MailingListSignupForm } from './MailingListSignup/MailingListSignup';
 import { Spacer } from './Spacer';
 import { MaxWidthContainer } from './MaxWidthContainer/MaxWidthContainer';
 import { useCurrentUrl } from '../hooks/useCurrentUrl';
